% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_function.R
\name{count_peaks}
\alias{count_peaks}
\title{Count number of paired insertions in each peak}
\usage{
count_peaks(
  peak_sets,
  filtered_fragments,
  extend_size,
  n_features = length(peak_sets)
)
}
\arguments{
\item{peak_sets}{GRanges object of peak sets}

\item{filtered_fragments}{filtered fragment also as a GRanges object}

\item{extend_size}{How long should we extend the exact insertion
site as accessible window}

\item{n_features}{Number of features (peaks)}
}
\value{
A sparse vector of PIC for each peak
}
\description{
Count number of paired insertions in each peak
}
