% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{.get_theoretical_c1}
\alias{.get_theoretical_c1}
\title{Title get theoretical snATAC-seq distribution under condition 1}
\usage{
.get_theoretical_c1(insertion_rate, peak_length = 500)
}
\arguments{
\item{insertion_rate}{The insertion rate (per base pair)}

\item{peak_length}{The width of peak}
}
\value{
A vector of length 6, representing the probability of observing 0 to >=5 counts
}
\description{
Title get theoretical snATAC-seq distribution under condition 1
}
\keyword{internal}
