% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{.get_theoretical_c12}
\alias{.get_theoretical_c12}
\title{Title get_theoretical_c12}
\usage{
.get_theoretical_c12(
  insertion_rate,
  peak_length = 1000,
  pad_length = 0,
  dirs = c(-1, 1),
  count_type = "PIC",
  min_frag_length = 25,
  max_frag_length = 600,
  cap_insertion = 20
)
}
\arguments{
\item{insertion_rate}{A single value of insertion rate}

\item{peak_length}{A single vale of peak length, default 1000}

\item{pad_length}{Length of pad region, i.e., flanking regions of the peak
where insertion can also happen}

\item{dirs}{Directions of insertion, for standard ATAC experiment (default),
it is c(-1,1)}

\item{count_type}{output count type, default is 'PIC'}

\item{min_frag_length}{minimum fragment length, default 25}

\item{max_frag_length}{maximum fragment length, default 600}

\item{cap_insertion}{number of insertions capped at this value, default 20

return A vector of probability corresponding to different number of fragments
#' @noRd}
}
\description{
Title get_theoretical_c12
}
