% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{.insertion_to_c1}
\alias{.insertion_to_c1}
\title{Title Construct an insertion-rate by mean count matrix}
\usage{
.insertion_to_c1(
  insertion_rates = (1:2000) * 0.01 * 0.001,
  peak_lengths = 4:20 * 50
)
}
\arguments{
\item{insertion_rates}{A vector of insertion rates (per base pair)}

\item{peak_lengths}{A vector containing width of peaks}
}
\value{
An insertion-rate by mean count matrix
}
\description{
Title Construct an insertion-rate by mean count matrix
}
\keyword{internal}
