% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{.insertion_to_c12}
\alias{.insertion_to_c12}
\title{Title insertion_to_c12}
\usage{
.insertion_to_c12(
  insertion_rates = (1:2000) * 0.01,
  peak_lengths = 4:20 * 50,
  pad_length = 0,
  dirs = c(-1, 1),
  count_type = "fragment",
  min_frag_length = 25,
  max_frag_length = 600,
  cap_insertion = 20
)
}
\arguments{
\item{insertion_rates}{A vector of insertion rates,
note that the insertion rate is per base, so the default range
is from 0.005 to 10}

\item{peak_lengths}{A vector of peak length ranges, default is
from 200 to 1000}
}
\value{
A insertion rate by peak length matrix, and each
element of the matrix corresponds to the number of expected fragments
#' @noRd
}
\description{
Title insertion_to_c12
}
