% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{.log_loss_frag}
\alias{.log_loss_frag}
\title{Title Compute log loss for a given estimated insertion rate}
\usage{
.log_loss_frag(
  est_inser,
  peak_length,
  cap_fragment = 5,
  capturing_rates,
  obs_pic_vec
)
}
\arguments{
\item{est_inser}{Estimated insertion rate}

\item{peak_length}{The width of peak}

\item{cap_fragment}{Maximum possible fragment (PIC) counts, default = 5}

\item{capturing_rates}{A vector of capturing rates in each cell}

\item{obs_pic_vec}{A vector specifying observed fragment (PIC) counts}
}
\value{
Log loss value
}
\description{
Title Compute log loss for a given estimated insertion rate
}
\keyword{internal}
