% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_estimate_joint_p_q.R
\name{get_r_by_ct_mat_pq}
\alias{get_r_by_ct_mat_pq}
\title{Get region (peak) by cell type matrix and per cell capturing rate}
\usage{
get_r_by_ct_mat_pq(
  cell_type_set,
  r_by_c,
  cell_type_labels,
  n_features_per_cell = nrow(r_by_c),
  p_acc = 5e-04,
  q_acc = 5e-04,
  n_max_iter = 800,
  verbose = TRUE
)
}
\arguments{
\item{cell_type_set}{A vector containing all cell types}

\item{r_by_c}{Input region (peak) by cell matrix. Nonzero entries are
binarized internally.}

\item{cell_type_labels}{A vector containing cell type labels}

\item{n_features_per_cell}{The number of features in the matrix,
defaulting to \code{nrow(r_by_c)}. If supplied, it must match that value.}

\item{p_acc}{The accuracy of p, default specified as 0.0005}

\item{q_acc}{The accuracy of q, default specified as 0.0005}

\item{n_max_iter}{The maximum iteration, default = 800}

\item{verbose}{Whether to output information on processing status}
}
\value{
A list with two elements, \itemize{
\item p_by_t Peak by cell type matrix, each element represents
the open probability of the peak in the corresponding cell type
\item q_vec A vector of cell-specific capturing rate
}
}
\description{
For a snATAC-seq (binary) dataset, compute the peak-specific
open probability
and cell-specific capturing rates
}
