% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{obs_to_insertion_MLE}
\alias{obs_to_insertion_MLE}
\title{Title Calculate the insertion rate from observed PIC counts}
\usage{
obs_to_insertion_MLE(pic_mat, capturing_rates, plen)
}
\arguments{
\item{pic_mat}{The observed peak by cell PIC count matrix}

\item{capturing_rates}{A vector of estimated capturing rates for each cell}

\item{plen}{A vector of peak widths}
}
\value{
The optimized insertion rate
}
\description{
Title Calculate the insertion rate from observed PIC counts
}
