% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{obs_to_insertion_MLE_lam}
\alias{obs_to_insertion_MLE_lam}
\title{Calculate insertion rates from observed PIC counts}
\usage{
obs_to_insertion_MLE_lam(pic_mat, capturing_rates, plen, n_cores = 1L)
}
\arguments{
\item{pic_mat}{The observed peak by cell PIC count matrix}

\item{capturing_rates}{A vector of estimated capturing rates for each cell}

\item{plen}{A vector of peak widths}

\item{n_cores}{A positive integer specifying the number of cores. On Windows,
values greater than one fall back to serial evaluation.}
}
\value{
The optimized insertion rate (per 1000 base pairs) over
insertion rates from 0.01 to 20
}
\description{
Calculate insertion rates from observed PIC counts
}
