% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_function.R
\name{PIC_counting}
\alias{PIC_counting}
\title{Count snATAC-seq data matrix with Paired Insertion Counting (PIC)}
\usage{
PIC_counting(
  cells,
  fragment_tsv_gz_file_location,
  peak_sets,
  deduplicate = FALSE,
  load_full = TRUE,
  extend_size = 5L,
  verbose = TRUE
)
}
\arguments{
\item{cells}{Cell barcode labels as a character vector.}

\item{fragment_tsv_gz_file_location}{The 10X Cell Ranger output
fragment.tsv.gz file location. This can usually be found at the /out
directory from Cell Ranger output}

\item{peak_sets}{The set of peaks as a GenomicRanges object. This will be
the features for the data matrix. Alternatively, this can be a data.frame
and the function will convert it into a GenomicRanges object}

\item{deduplicate}{Whether to include deduplicate step where within
the same cell,
fragments with identical start and end location will be deduplicated.
This is usually unnecessary for Cell Ranger ATAC output, since
Cell Ranger ATAC has already deduplicated the fragments.
But for dsc-ATAC-seq data, this step will
be helpful and recommended.}

\item{load_full}{Whether to load the whole fragment.tsv.gz file into memory.
If set to \code{FALSE}, the function loads it by chromosome to save RAM. This
mode requires a block-gzipped file and its Tabix index (\code{.tbi}).}

\item{extend_size}{How long should we extend the exact insertion site as
an accessible window, in base pairs.}

\item{verbose}{Whether to output progress information including the progress
bar}
}
\value{
A sparse peak-by-cell PIC count matrix.
}
\description{
Count snATAC-seq data matrix with Paired Insertion Counting (PIC)
}
