% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{DAR_by_LRT}
\alias{DAR_by_LRT}
\title{Compute p value for DAR test between two cell types}
\usage{
DAR_by_LRT(
  pic_mat,
  capturing_rates,
  cell_type_labels,
  n_cores = 1,
  plen = NULL,
  min_frag_length = 25,
  max_frag_length = 600,
  estimation_approach = "MLE"
)
}
\arguments{
\item{pic_mat}{The observed peak by cell PIC count matrix}

\item{capturing_rates}{A vector of estimated capturing rates for each cell}

\item{cell_type_labels}{A vector specifying cell type labels}

\item{n_cores}{A numerical value to specify the number of cores in parallel,
default = 1}

\item{plen}{A vector of peak length}

\item{min_frag_length}{The value for the s1 hyperparameter in the ssPoisson
distribution, this stands for the minimum fragment length requirement such
that the fragment can be amplifiable and mappable to genome. Default = 25}

\item{max_frag_length}{The value for the s2 hyperparameter in the ssPoisson
distribution, this stands for the max fragment length requirement such
that the fragment can be amplifiable. Default = 600}

\item{estimation_approach}{The approach for parameter estimation,
either 'MLE'
for condition 1 or 'ME' for condition 1+2.
The 'MLE' approach is more accurate
and usually it has a higher power, but it ignores the size filtering step
in snATAC-seq data generation. Default is 'MLE'}
}
\value{
Log loss value
}
\description{
Compute p value for DAR test between two cell types
}
