% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{DAR_by_LRT}
\alias{DAR_by_LRT}
\title{Title Compute p value for DAR test between two cell types}
\usage{
DAR_by_LRT(pic_mat, capturing_rates, cell_type_labels)
}
\arguments{
\item{pic_mat}{The observed peak by cell PIC count matrix}

\item{capturing_rates}{A vector of estimated capturing rates for each cell}

\item{cell_type_labels}{A vector specifying cell type labels}
}
\value{
Log loss value
}
\description{
Title Compute p value for DAR test between two cell types
}
