% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_function.R
\name{load_fragments}
\alias{load_fragments}
\title{Function for loading fragment files and filter by cell barcodes}
\usage{
load_fragments(fragment_tsv_gz_file_location, cells, verbose = TRUE)
}
\arguments{
\item{fragment_tsv_gz_file_location}{The 10X Cell Ranger output
fragment.tsv.gz file location. This can usually be found at the /out
directory from Cell Ranger output}

\item{cells}{The cell barcode lables as a Character vector}

\item{verbose}{Whether to output progress message. Default TRUE}
}
\value{
data.frame containing fragments that are filtered by cell barcodes
}
\description{
Function for loading fragment files and filter by cell barcodes
}
