% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insertion_rates_and_DAR_test.R
\name{obs_to_insertion_MLE_lam}
\alias{obs_to_insertion_MLE_lam}
\title{Calculate the insertion rate from observed PIC counts}
\usage{
obs_to_insertion_MLE_lam(pic_mat, capturing_rates, plen, n_cores)
}
\arguments{
\item{pic_mat}{The observed peak by cell PIC count matrix}

\item{capturing_rates}{A vector of estimated capturing rates for each cell}

\item{plen}{A vector of peak widths}

\item{n_cores}{A numerical value to specify the number of cores in parallel}
}
\value{
The optimized insertion rate
}
\description{
Calculate the insertion rate from observed PIC counts
}
