% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_function.R
\name{PIC_counting}
\alias{PIC_counting}
\title{Title PIC-count}
\usage{
PIC_counting(cells, fragment_tsv_gz_file_location, peak_sets)
}
\arguments{
\item{cells}{The cell barcode lables as a Character vector}

\item{fragment_tsv_gz_file_location}{The 10X Cell Ranger output fragment.tsv.gz file location}

\item{peak_sets}{The set of peaks as a GenomicRanges object}
}
\value{
The counted cell by peak matrix
}
\description{
Title PIC-count
}
